+++ authpass/lib/utils/error_utils.dart
@@ -1,0 +1,3 @@
+Never throws(Object e) {
+  throw e;
+}